fix: emit noindex for Forbidden/Unauthorized boundaries, extract retryScrollTo helper
